some architecture changes
