+++ src/components/lightDarkTheme/LightDarkTheme.tsx
@@ -1,28 +1,8 @@
-import { useState, useEffect, useCallback } from 'react';
+import { useLightDarkTheme } from './hook';
 
 export const LightDarkThemeToggle = () => {
-  const defaultTheme = window.matchMedia('(prefers-color-scheme: dark)')?.matches ? 'dark' : 'light';
-  const [theme, setTheme] = useState(localStorage.getItem('theme') || defaultTheme);
-
-  const toggleTheme = useCallback(() => {
-    const newTheme = theme === 'light' ? 'dark' : 'light';
-    setTheme(newTheme);
-    localStorage.setItem('theme', newTheme);
-  }, [theme]);
-
-  useEffect(() => {
-    applyThemeStyles();
-  }, [theme]);
-
-  const applyThemeStyles = () => {
-    const body = document.body;
-    if (theme === 'light') {
-      body.setAttribute('data-theme', 'light');
-    } else {
-      body.setAttribute('data-theme', 'dark');
-    }
-  };
-
+  const { toggleTheme } = useLightDarkTheme()
+  
   return (
     <div>
       <button onClick={toggleTheme}>ToggleTheme</button>
