+++ src/constants/lightDarkTheme.ts
@@ -1,0 +1,2 @@
+export const dark = 'dark';
+export const light = 'light';
